feat(citation): move citation fuzzy matching code and be less intrusive
